requirements.txt
#python==3.12.4
attrs==25.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
duckdb==1.1.3
elastic-transport==8.15.1
elasticsearch==8.16.0
et_xmlfile==2.0.0
fonttools==4.55.8
geopandas==1.0.1
h11==0.14.0
h3==4.2.1
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.1.3
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyarrow==18.1.0
pycparser==2.22
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
selenium==4.28.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
